*: subjectManifest -> subject
the ability to establish a reference between the object just pushed to
the registry, and an existing object, ought not be limited to _only_
referencing a "manifest".

The use case here is that: as a builder of a base image, I may push a
signature or SBOM, referencing the digest of that image's _manifest_,
*but* as soon as consumer of that image build's FROM it, then the only
digest certain to persist is the layer(s). Allowing a `subject` to be
the layer itself could allow the final consumer of a constructed image
to attest to _all_ the layers and the final image itself.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>